'Implemented job 6'
Each question in job 6 has been written in one file, as well as each to
its own file. These scripts print the current table number being run
(for gauging progress) and print as the last number the number of
contradictions found. As of this commit, no contradictions have been
found for semigroups of order 4 or less.